Move error callstack to RPC-only interface
- Add IChatAgentInvokeResult to extHost.protocol.ts with errorCallstack
- Emit chatAgentError telemetry at the RPC boundary in mainThreadChatAgents2
- Strip errorCallstack before returning to model layer
- Remove callstack from IChatResponseErrorDetails (no longer needed)
- Remove callstack telemetry handling from chatServiceImpl.ts
