+++ src/vs/workbench/contrib/chat/common/chatService/chatServiceImpl.ts
@@ -27,7 +27,6 @@ import { ILogService } from '../../../../../platform/log/common/log.js';
 import { Progress } from '../../../../../platform/progress/common/progress.js';
 import { IStorageService, StorageScope, StorageTarget } from '../../../../../platform/storage/common/storage.js';
 import { ITelemetryService } from '../../../../../platform/telemetry/common/telemetry.js';
-import { packErrorForTelemetry } from '../../../../../platform/telemetry/common/errorTelemetry.js';
 import { IWorkspaceContextService } from '../../../../../platform/workspace/common/workspace.js';
 import { IExtensionService } from '../../../../services/extensions/common/extensions.js';
 import { IChatEntitlementService } from '../../../../services/chat/common/chatEntitlementService.js';
@@ -1353,12 +1352,6 @@ export class ChatService extends Disposable implements IChatService {
 						request,
 					});
 
-					if (rawResult.errorDetails?.callstack) {
-						const errorAgent = detectedAgent ?? agentPart?.agent ?? defaultAgent;
-						logChatAgentError(this.telemetryService, errorAgent, rawResult.errorDetails.callstack, rawResult.errorDetails.message);
-						delete rawResult.errorDetails.callstack;
-					}
-
 					model.setResponse(request, rawResult);
 					completeResponseCreated();
 					this.trace('sendRequest', `Provider returned response for session ${model.sessionResource}`);
@@ -1389,11 +1382,6 @@ export class ChatService extends Disposable implements IChatService {
 						detectedAgent,
 						request,
 					});
-					const { callstack, msg } = packErrorForTelemetry(err);
-					if (callstack) {
-						const errorAgent = detectedAgent ?? agentPart?.agent ?? defaultAgent;
-						logChatAgentError(this.telemetryService, errorAgent, callstack, msg);
-					}
 					const rawResult: IChatAgentResult = { errorDetails: { message: err.message } };
 					model.setResponse(request, rawResult);
 					completeResponseCreated();
@@ -1847,22 +1835,4 @@ export async function chatModelToChatDetail(model: IChatModel): Promise<IChatDet
 		stats: await awaitStatsForSession(model),
 		lastResponseState: model.lastRequest?.response?.state ?? ResponseModelState.Pending,
 	};
-}
-
-function logChatAgentError(telemetryService: ITelemetryService, agent: IChatAgentData | undefined, callstack: string, msg: string): void {
-	type ChatAgentErrorEvent = { callstack: string; msg: string; agent: string; agentExtensionId: string };
-	type ChatAgentErrorClassification = {
-		owner: 'bryanchen-d';
-		comment: 'Logged when a chat agent handler throws an error with a callstack.';
-		callstack: { classification: 'CallstackOrException'; purpose: 'PerformanceAndHealth'; comment: 'The callstack of the error.' };
-		msg: { classification: 'CallstackOrException'; purpose: 'PerformanceAndHealth'; comment: 'The error message.' };
-		agent: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The agent that threw the error.' };
-		agentExtensionId: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The extension that contributed the agent.' };
-	};
-	telemetryService.publicLogError2<ChatAgentErrorEvent, ChatAgentErrorClassification>('chatAgentError', {
-		callstack,
-		msg,
-		agent: agent?.id ?? '',
-		agentExtensionId: agent?.extensionId.value ?? '',
-	});
 }
